requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework-simplejwt==5.2.0
djangorestframework==3.13.1
PyJWT==2.4.0
pytz==2022.1
setuptools==63.3.0
sqlparse==0.4.2
wheel==0.37.1
